addressed review comments in Integration tests
