+++ lom/integration_test/src/kill.py
@@ -3,29 +3,24 @@ import sys
 
 import api
 
-# Function to kill LoMEngine
-def kill_lomengine():
-    engine_pids = api.get_lomengine_pids()
-    if engine_pids:
-        print("Killing LoMEngine instances...")
-        for pid in engine_pids:
-            print(f"Killing LoMEngine (PID: {pid})...")
-            subprocess.run(['sudo', "kill", str(pid)])
-        print("LoMEngine instances killed.")
+# Function to kill processes by name
+def kill_processes_by_name(process_name):
+    if process_name == "LoMEngine":
+        process_pids = api.get_lomengine_pids()
+    elif process_name == "LoMPluginMgr":
+        process_pids = api.get_lompluginmgr_pids()
     else:
-        print("No LoMEngine instances are running.")
-
-# Function to kill all LoMPluginMgr instances
-def kill_lompluginmgr():
-    pluginmgr_pids = api.get_lompluginmgr_pids()
-    if pluginmgr_pids:
-        print("Killing LoMPluginMgr instances...")
-        for pid in pluginmgr_pids:
-            print(f"Killing LoMPluginMgr (PID: {pid})...")
+        print(f"Error: Unknown process name '{process_name}'")
+        return
+    
+    if process_pids:
+        print(f"Killing {process_name} instances...")
+        for pid in process_pids:
+            print(f"Killing {process_name} (PID: {pid})...")
             subprocess.run(['sudo', "kill", str(pid)])
-        print("LoMPluginMgr instances killed.")
+        print(f"{process_name} instances killed.")
     else:
-        print("No LoMPluginMgr instances are running.")
+        print(f"No {process_name} instances are running.")
 
 # Print usage information
 def print_usage():
@@ -42,12 +37,12 @@ if __name__ == '__main__':
 
     arg = sys.argv[1]
     if arg == "all":
-        kill_lomengine()
-        kill_lompluginmgr()
+        kill_processes_by_name("LoMEngine")
+        kill_processes_by_name("LoMPluginMgr")
     elif arg == "engine":
-        kill_lomengine()
+        kill_processes_by_name("LoMEngine")
     elif arg == "plugin":
-        kill_lompluginmgr()
+        kill_processes_by_name("LoMPluginMgr")
     else:
         print_usage()
         exit(1)
