+++ lom/integration_test/tests/Test_12_Heartbeats.py
@@ -5,55 +5,66 @@ import re
 import time
 import select
 import contextlib
+from datetime import datetime
+
 import src.api as api
 
 def isMandatoryPass() :
     return True
 
 def getTestName() :
-    return "link_crc detection Only first and last outlier  1,0,0,0, 1,0,0,0  "
+    return "Test heartbeats from plugin to engine"
 
 def getTestDescription() :
-    return "link_crc detection Only first and last outlier  \
-            pattern 1,0,0,0, 1,0,0,0   \
-            Minimum time for detection = 150 sec  \
-            Maximum time for detection = 240 sec \
-            "
+    return " Plugin must send heartbeats to engine periodically. Test that period \
+            When anomaly is detected, again when new request is processed, heartbeats must resume \
+        "
 
-def enableTest() :
-    return False
+def isEnabled() :
+    return False 
 
 def run_test():
-    
-    # Specify the patterns to be matched for engine and plmgr syslogs
+
+    # Specify the patterns to be matched/not matched for engine logs
     engine_pat1 = r"{\"LoM_Action\":{\"Action\":\"link_crc\",\"InstanceId\":\"([\w-]+)\",\"AnomalyInstanceId\":\"([\w-]+)\",\"AnomalyKey\":\"(\w+)\",\"Response\":\"Detected Crc\",\"ResultCode\":(\d+),\"ResultStr\":\"Success\"},\"State\":\"init\"}"
     engine_pat2 = r"{\"LoM_Action\":{\"Action\":\"link_crc\",\"InstanceId\":\"([\w-]+)\",\"AnomalyInstanceId\":\"([\w-]+)\",\"AnomalyKey\":\"(\w+)\",\"Response\":\"Detected Crc\",\"ResultCode\":(\d+),\"ResultStr\":\"No follow up actions \(seq:link_crc_bind-0\)\"},\"State\":\"complete\"}"
 
     engine_patterns = [
-        (api.PATTERN_MATCH, engine_pat1),
-        (api.PATTERN_MATCH, engine_pat2)
+          (api.PATTERN_MATCH, engine_pat1), # This pattern must match
+          (api.PATTERN_MATCH, engine_pat2) # This pattern must match
     ]
     
-    # Specify the patterns to be matched for plmgr syslogs
+    # Specify the patterns to be matched/not matched for plmgr syslogs 
     plugin_pat_1 = "link_crc: executeCrcDetection Anomaly Detected"
     plugin_pat_2 = r"In handleRequest\(\): Received response from plugin link_crc, data : Action: link_crc InstanceId: ([\w-]+) AnomalyInstanceId: ([\w-]+) AnomalyKey: (\w+) Response: Detected Crc ResultCode: (\d+) ResultStr: Success"
     plugin_pat_3 = r"In handleRequest\(\): Completed processing action request for plugin:link_crc"
     plugin_pat_4 = r"In run\(\) : Sending response to engine : Action: link_crc InstanceId: ([\w-]+) AnomalyInstanceId: ([\w-]+) AnomalyKey: (\w+) Response: Detected Crc ResultCode: (\d+) ResultStr: Success"
     plugin_pat_5 = r"SendServerResponse: succeeded \(proc_0/RecvServerRequestAction\)"
-    plugin_pat_6 = r"link_crc: ExecuteCrcDetection Starting"
+    plugin_pat_6 = r"RecvServerRequest: succeeded \(proc_0/RecvServerRequestAction\)"
+    plugin_pat_7 = r"In run\(\) RecvServerRequest : Received action request : Action: (\w+) InstanceId: ([\w-]+) AnomalyInstanceId: ([\w-]+) AnomalyKey:  Timeout: (\d+)"
+    plugin_pat_8 = r"In handleRequest\(\): Processing action request for plugin:link_crc, timeout:(\d+) InstanceId:([\w-]+) AnomalyInstanceId:([\w-]+) AnomalyKey:"
+    plugin_pat_9 = r"STarted Request\(\) for \((\w+)\)"
+    plugin_pat_10 = r"link_crc: ExecuteCrcDetection Starting"
+    plugin_pat_11 = r"In run\(\) : Sending response to engine : Heartbeat notification: link_crc (\d+)"
+
 
     plmgr_patterns = [
-        (api.PATTERN_MATCH, plugin_pat_1),
-        (api.PATTERN_MATCH, plugin_pat_2),
-        (api.PATTERN_MATCH, plugin_pat_3),
-        (api.PATTERN_MATCH, plugin_pat_4),
-        (api.PATTERN_MATCH, plugin_pat_5),
-        (api.PATTERN_MATCH, plugin_pat_6)
+          (api.PATTERN_MATCH, plugin_pat_1), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_2), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_3), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_4), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_5), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_6), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_7), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_8), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_9), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_10), # This pattern must match
+          (api.PATTERN_MATCH, plugin_pat_11) # This pattern must match
     ]
 
     # Specify the minimum and maximum detection time in seconds
-    MIN_DETECTION_TIME = 150
-    MAX_DETECTION_TIME = 240
+    MIN_DETECTION_TIME = 60
+    MAX_DETECTION_TIME = 60
 
     # Overwrite config files with test specific data in Docker container
     json_data = {
@@ -106,7 +117,7 @@ def run_test():
         "MAX_SEQ_TIMEOUT_SECS": 120,
         "MIN_PERIODIC_LOG_PERIOD_SECS": 1,
         "ENGINE_HB_INTERVAL_SECS": 10,
-        "INITIAL_DETECTION_REPORTING_FREQ_IN_MINS": 5,
+        "INITIAL_DETECTION_REPORTING_FREQ_IN_MINS": 5, 
         "SUBSEQUENT_DETECTION_REPORTING_FREQ_IN_MINS": 60,
         "INITIAL_DETECTION_REPORTING_MAX_COUNT": 12,
         "PLUGIN_MIN_ERR_CNT_TO_SKIP_HEARTBEAT": 3,
@@ -150,7 +161,7 @@ def run_test():
     log_monitor = api.LogMonitor()
 
     # Create an instance of BinaryRunner
-    binary_runner = api.BinaryRunner("../bin/linkcrc_mocker", "2", "Ethernet96")
+    binary_runner = api.BinaryRunner("../bin/linkcrc_mocker", "0", "Ethernet96")
 
     # Create separate events to signal the monitoring threads to stop
     engine_stop_event = threading.Event()
@@ -159,20 +170,21 @@ def run_test():
     # Create a list to hold the monitoring threads
     monitor_threads = []
 
-    # Start the syslog monitoring thread for engine syslogs
-    monitor_engine_thread_1 = threading.Thread(target=log_monitor.monitor_engine_syslogs_noblock, args=(engine_patterns, engine_stop_event))
+    # Start the syslog monitoring thread for engine syslogs. Force wait untill monitoring_duration is expired as need to match all dublicate logs
+    monitor_engine_thread_1 = threading.Thread(target=log_monitor.monitor_engine_syslogs_noblock, args=(engine_patterns, engine_stop_event, True))
     monitor_engine_thread_1.start()
     monitor_threads.append(monitor_engine_thread_1)
 
-    # Start the syslog monitoring thread for plmgr syslogs
-    monitor_plmgr_thread_1 = threading.Thread(target=log_monitor.monitor_plmgr_syslogs_noblock, args=(plmgr_patterns, plmgr_instance, plmgr_stop_event))
+    # Start the syslog monitoring thread for plmgr syslogs. Force wait untill monitoring_duration is expired as need to match all dublicate logs
+    monitor_plmgr_thread_1 = threading.Thread(target=log_monitor.monitor_plmgr_syslogs_noblock, args=(plmgr_patterns, plmgr_instance, plmgr_stop_event, True))
     monitor_plmgr_thread_1.start()
     monitor_threads.append(monitor_plmgr_thread_1)
 
-    with monitor_threads_context(monitor_threads, engine_stop_event, plmgr_stop_event): 
+    with monitor_threads_context(monitor_threads, engine_stop_event, plmgr_stop_event):   
             
         # Stop the binary
-        binary_runner.stop_binary()
+        if binary_runner.stop_binary() == False:
+            return api.TEST_FAIL
 
         # Start the binary
         if not binary_runner.run_binary_in_background():
@@ -181,31 +193,33 @@ def run_test():
             return api.TEST_FAIL
         print("linkcrc_mocker started ...........")
 
-        # Wait for a specified duration to monitor the logs(2 outliers)
-        monitoring_duration = MAX_DETECTION_TIME + 15  # Specify the duration in seconds
+        # Wait for a specified duration to monitor the logs(one outlier)
+        monitoring_duration = MAX_DETECTION_TIME + 60  # Specify the duration in seconds
         time.sleep(monitoring_duration)
 
-        # Stop the monitoring threads and join them
+         # Stop the monitoring threads and join them
         stop_and_join_threads(monitor_threads, engine_stop_event, plmgr_stop_event)
 
         # Stop the binary
         if not binary_runner.stop_binary():
             print("Failed to stop linkcrc_mocker")
-            return 1
-
+            return api.TEST_FAIL
+        
         # Determine the test results based on the matched patterns
         status = api.TEST_PASS  # Return code 0 for test success
 
+        ########## check that engine logs must contain all patterns in engine_patterns of type PATTERN_MATCH
         engine_match_count = 0
         for flag, pattern in engine_patterns:
             if flag == api.PATTERN_MATCH:
-                if pattern in log_monitor.engine_matched_patterns:
-                    engine_match_count += 1
-                    print(f"\nExpected, Matched engine pattern ------------------ \n'{pattern}' \nMatch Message ------------------")
-                    for timestamp, log_message in log_monitor.engine_matched_patterns.get(pattern, []):
-                        print(f"Timestamp: {timestamp}, Log Message: {log_message}")
-                else:
-                    print(f"\nUnExpected, No match found for engine pattern ------------------ '{pattern}'")
+                with log_monitor.match_lock:  # Added lock here
+                    if pattern in log_monitor.engine_matched_patterns:
+                        engine_match_count += 1
+                        print(f"\nExpected, Matched engine pattern ------------------ \n'{pattern}' \nMatch Message ------------------")
+                        for timestamp, log_message in log_monitor.engine_matched_patterns.get(pattern, []):
+                            print(f"Timestamp: {timestamp}, Log Message: {log_message}")
+                    else:
+                        print(f"\nUnExpected, No match found for engine pattern ------------------ '{pattern}'")
 
         expected_engine_match_count = len([p for t, p in engine_patterns if t == api.PATTERN_MATCH])
         if engine_match_count == expected_engine_match_count:
@@ -213,18 +227,19 @@ def run_test():
         else:
             print(f"\nFail, Expected engine match count: {expected_engine_match_count}, Actual count: {engine_match_count}. Some engine match patterns not matched for Test Case. Test for engine failed.")
             status = api.TEST_FAIL  # Return code 1 for test failure
-                    
-        # check that plmgr logs must contain all patterns in plmgr_patterns of type PATTERN_MATCH
+
+        ########### check that plmgr logs must contain all patterns in plmgr_patterns of type PATTERN_MATCH
         plmgr_match_count = 0
         for flag, pattern in plmgr_patterns:
             if flag == api.PATTERN_MATCH:
-                if pattern in log_monitor.plmgr_matched_patterns:
-                    plmgr_match_count += 1
-                    print(f"\nExpected, Matched Plmgr pattern ------------------ \n'{pattern}' \nMatch Message ------------------")
-                    for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(pattern, []):
-                        print(f"Timestamp: {timestamp}, Log Message: {log_message}")
-                else:
-                    print(f"\nUnExpected, No match found for plmgr pattern ------------------ '{pattern}'")
+                with log_monitor.match_lock:  # Added lock here
+                    if pattern in log_monitor.plmgr_matched_patterns:
+                        plmgr_match_count += 1
+                        print(f"\nExpected, Matched Plmgr pattern ------------------ \n'{pattern}' \nMatch Message ------------------")
+                        for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(pattern, []):
+                            print(f"Timestamp: {timestamp}, Log Message: {log_message}")
+                    else:
+                        print(f"\nUnExpected, No match found for plmgr pattern ------------------ '{pattern}'")
 
         expected_plmgr_match_count = len([p for t, p in plmgr_patterns if t == api.PATTERN_MATCH])
         if plmgr_match_count == expected_plmgr_match_count:
@@ -233,53 +248,56 @@ def run_test():
             print(f"\nFail, Expected PLMGR match count: {expected_plmgr_match_count}, Actual count: {plmgr_match_count}. Some PLMGR match patterns not matched for Test Case. Test for PLMGR failed.")
             status = api.TEST_FAIL  # Return code 1 for test failure
 
-        
-        # Get the InstanceId, AnomalyInstanceId and AnomalyKey from the plugin manager logs
-        InstanceId = None
-        AnomalyInstanceId = None
-        AnomalyKey = None
-        
-        for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(plugin_pat_2, []):
-            #print(f"Timestamp::: {timestamp}, Log Message::: {log_message}")
-            match = re.search(plugin_pat_2, log_message)
-            if match:
-                InstanceId = match.group(1)
-                AnomalyInstanceId = match.group(2)
-                AnomalyKey = match.group(3)
-                print(f"SUccess : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} from plmgr logs")
-                break    
-        if InstanceId is None or AnomalyInstanceId is None or AnomalyKey is None:
-            print(f"Fail : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} not found in PLMGR logs")
-            status = api.TEST_FAIL
-        
-        # cross check the above InstanceId, AnomalyInstanceId and AnomalyKey with the engine logs 
-        if InstanceId is not None and AnomalyInstanceId is not None and AnomalyKey is not None:
-            for timestamp, log_message in log_monitor.engine_matched_patterns.get(engine_pat1, []):
-                #print(f"Timestamp::: {timestamp}, Log Message::: {log_message}")
-                match = re.search(engine_pat1, log_message)
-                if match:
-                    if InstanceId == match.group(1) and AnomalyInstanceId == match.group(2) and AnomalyKey == match.group(3):
-                        print(f"Success : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} matched with engine logs")
-                        break
-                    else:
-                        print(f"Fail : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} not matched with engine logs")
-                        status = api.TEST_FAIL
-                        break
+        if status == api.TEST_FAIL:
+            print("Fail : Unable to proceed test further due to above failed conditions")
+            return status
 
-        # Cross check the above InstanceId, AnomalyInstanceId and AnomalyKey with the next sequence of engine logs 
-        if InstanceId is not None and AnomalyInstanceId is not None and AnomalyKey is not None:
-            for timestamp, log_message in log_monitor.engine_matched_patterns.get(engine_pat2, []):
-                #print(f"Timestamp::: {timestamp}, Log Message::: {log_message}")
-                match = re.search(engine_pat2, log_message)
-                if match:
-                    if InstanceId == match.group(1) and AnomalyInstanceId == match.group(2) and AnomalyKey == match.group(3):
-                        print(f"Success : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} matched with engine logs")
-                        break
-                    else:
-                        print(f"Fail : InstanceId: {InstanceId}, AnomalyInstanceId: {AnomalyInstanceId}, AnomalyKey: {AnomalyKey} not matched with engine logs")
-                        status = api.TEST_FAIL
-                        break
+        ######### Chech the heartbeats from plugin i.e. plugin_pat_11. Each heartbeat must be send periodically every 30 sec
+        print(f"\nChecking heartbeats are sent periodically or not from plugin\n")
+        with log_monitor.match_lock:  # Added lock here
+            if plugin_pat_11 in log_monitor.plmgr_matched_patterns:
+                previousTimestamp = None                
+                for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(plugin_pat_11, []):
+                    print(f"Timestamp: {timestamp}, Log Message: {log_message}")
+                    if previousTimestamp is not None:                        
+                        if (datetime.strptime(timestamp, "%b %d %H:%M:%S.%f") - datetime.strptime(previousTimestamp, "%b %d %H:%M:%S.%f")).total_seconds() >= 30 and \
+                                (datetime.strptime(timestamp, "%b %d %H:%M:%S.%f") - datetime.strptime(previousTimestamp, "%b %d %H:%M:%S.%f")).total_seconds() <= 31:
+                            print(f"\nSuccess, Heartbeats are sent periodically from plugin")
+                            previousTimestamp = timestamp
+                        else:
+                            print(f"\nFail, Heartbeats are not sent periodically from plugin")
+                            status = api.TEST_FAIL
+        
+        # Now to make sure after anomaly is detected, heartbeats must resume and check that heartbeats are sent periodically or not from plugin
+        
+        print(f"\nChecking heartbeats are sent periodically or not from plugin after anomaly is detected\n")
+        
+        # Get timestamp of detection(plugin_pat_1)
+        timestamp_of_detection = None
+        if plugin_pat_1 in log_monitor.plmgr_matched_patterns:
+            for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(plugin_pat_1, []):
+                timestamp_of_detection = timestamp
+                break
+
+        with log_monitor.match_lock:  # Added lock here
+            if plugin_pat_11 in log_monitor.plmgr_matched_patterns:
+                previousTimestamp = None                
+                for timestamp, log_message in log_monitor.plmgr_matched_patterns.get(plugin_pat_11, []):
+                    if timestamp < timestamp_of_detection:
+                        continue
+                    print(f"Timestamp: {timestamp}, Log Message: {log_message}")
+                    if previousTimestamp is not None:                        
+                        if (datetime.strptime(timestamp, "%b %d %H:%M:%S.%f") - datetime.strptime(previousTimestamp, "%b %d %H:%M:%S.%f")).total_seconds() >= 30 and \
+                                (datetime.strptime(timestamp, "%b %d %H:%M:%S.%f") - datetime.strptime(previousTimestamp, "%b %d %H:%M:%S.%f")).total_seconds() <= 31 :
+                            print(f"\nSuccess, Heartbeats are sent periodically from plugin after anomaly is detected")
+                            previousTimestamp = timestamp
+                        else:
+                            print(f"\nFail, Heartbeats are not sent periodically from plugin after anomaly is detected")
+                            status = api.TEST_FAIL
+       
+
     return status
+
 
 def stop_and_join_threads(threads, engine_stop_event, plmgr_stop_event):
     # Set the event to stop the monitoring threads
