+++ lom/integration_test/tests/Test_1_SwitchBasic.py
@@ -8,9 +8,9 @@ def getTestName() :
     return "show version, check docker-device-health docker image"
 
 def getTestDescription() :
-    return "show version, check docker image"
+    return "show version of sonic build, check docker device-health image is installed or not"
 
-def enableTest() :
+def isEnabled() :
     return False
 
 
@@ -21,7 +21,7 @@ def run_test():
         print(output[1])
     else:
         if output[0] == "COMMAND_NOT_FOUND":
-            print("'show version' command not found")
+            print("'show version' command not found which is unexpected")
         elif output[0] == "ERROR":
             print("Error occurred while running 'show version' command.")
         return 1
@@ -36,8 +36,8 @@ def run_test():
         return 0
     else:
         if output[0] == "COMMAND_NOT_FOUND":
-            print("'docker images docker-device-health' command not found")
+            print("'Unexpected. docker images docker-device-health' command not found")
         elif output[0] == "ERROR":
-            print("Error occurred while running 'docker images docker-device-health' command.")
+            print("Unexpected. Error occurred while running 'docker images docker-device-health' command.")
         return 1
