+++ lom/integration_test/tests/Test_4_EngineCrash.py
@@ -13,7 +13,7 @@ def getTestDescription() :
             stopped and restarted \
             "
 
-def enableTest() :
+def isEnabled() :
     return False
 
 def run_test():
